修复antd兼容性，日志详情自动滚动
